delete old generate code code and fix tests
