+++ api/src/main/java/org/pmiops/workbench/dataset/DataSetService.java
@@ -6,12 +6,15 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+
+import org.json.JSONObject;
 import org.pmiops.workbench.db.model.DbCohort;
 import org.pmiops.workbench.db.model.DbConceptSet;
 import org.pmiops.workbench.db.model.DbDataset;
 import org.pmiops.workbench.db.model.DbWorkspace;
 import org.pmiops.workbench.model.DataDictionaryEntry;
 import org.pmiops.workbench.model.DataSet;
+import org.pmiops.workbench.model.DataSetExportRequest;
 import org.pmiops.workbench.model.DataSetPreviewRequest;
 import org.pmiops.workbench.model.DataSetRequest;
 import org.pmiops.workbench.model.DomainValue;
@@ -19,6 +22,8 @@ import org.pmiops.workbench.model.KernelTypeEnum;
 import org.pmiops.workbench.model.ResourceType;
 
 public interface DataSetService {
+
+  DbDataset mustGetDbDataset(long workspaceId, long dataSetId);
 
   DataSet saveDataSet(DataSetRequest dataSetRequest, Long userId);
 
@@ -36,6 +41,8 @@ public interface DataSetService {
       String cdrVersionName,
       String qualifier,
       Map<String, QueryJobConfiguration> queryJobConfigurationMap);
+
+  List<String> generateCodeCells(DataSetExportRequest dataSetExportRequest, DbWorkspace dbWorkspace);
 
   DbDataset cloneDataSetToWorkspace(
       DbDataset fromDataSet,
@@ -65,4 +72,6 @@ public interface DataSetService {
   List<String> getPersonIdsWithWholeGenome(DbDataset dataSet);
 
   List<DomainValue> getValueListFromDomain(String domain);
+
+  void validateDataSetPreviewRequestResources(long workspaceId, DataSetPreviewRequest request);
 }
