+++ api/src/main/resources/workbench-api.yaml
@@ -2232,29 +2232,6 @@ paths:
           description: Successful copy
           schema:
             "$ref": "#/definitions/ConceptSet"
-  "/v1/workspaces/{workspaceNamespace}/{workspaceId}/data-set/generateCode/{kernelType}":
-    post:
-      tags:
-      - dataSet
-      description: 'Given a Data Set, return the SQL query built from it.'
-      operationId: generateCode
-      parameters:
-      - "$ref": "#/parameters/workspaceNamespace"
-      - "$ref": "#/parameters/workspaceId"
-      - in: path
-        name: kernelType
-        required: true
-        type: string
-      - in: body
-        name: dataSet
-        required: true
-        schema:
-          "$ref": "#/definitions/DataSetRequest"
-      responses:
-        200:
-          description: 'A SQL query for each domain in the Data Set'
-          schema:
-            "$ref": "#/definitions/DataSetCodeResponse"
   "/v1/workspaces/{workspaceNamespace}/{workspaceId}/data-sets":
     parameters:
     - "$ref": "#/parameters/workspaceNamespace"
